+++ pages/dev_multi_tenant/src/payload-types.ts
@@ -112,10 +112,10 @@ export interface Config {
   db: {
     defaultIDType: number;
   };
-  fallbackLocale: null;
+  fallbackLocale: ('false' | 'none' | 'null') | false | null | ('de' | 'en') | ('de' | 'en')[];
   globals: {};
   globalsSelect: {};
-  locale: null;
+  locale: 'de' | 'en';
   widgets: {
     collections: CollectionsWidget;
   };
@@ -157,6 +157,13 @@ export interface Page {
   breadcrumbs: Breadcrumbs;
   title: string;
   content: string;
+  meta: {
+    alternatePaths: {
+      hreflang: string;
+      path: string;
+      id?: string | null;
+    }[];
+  };
   updatedAt: string;
   createdAt: string;
   deletedAt?: string | null;
@@ -171,6 +178,8 @@ export interface Tenant {
   slug: string;
   name: string;
   websiteUrl: string;
+  primaryLocale: 'de' | 'en';
+  prefixAllLocales?: boolean | null;
   updatedAt: string;
   createdAt: string;
 }
@@ -442,6 +451,17 @@ export interface PagesSelect<T extends boolean = true> {
   breadcrumbs?: T | BreadcrumbsSelect<T>;
   title?: T;
   content?: T;
+  meta?:
+    | T
+    | {
+        alternatePaths?:
+          | T
+          | {
+              hreflang?: T;
+              path?: T;
+              id?: T;
+            };
+      };
   updatedAt?: T;
   createdAt?: T;
   deletedAt?: T;
@@ -568,6 +588,8 @@ export interface TenantsSelect<T extends boolean = true> {
   slug?: T;
   name?: T;
   websiteUrl?: T;
+  primaryLocale?: T;
+  prefixAllLocales?: T;
   updatedAt?: T;
   createdAt?: T;
 }