+++ pages/src/exports/server.ts
@@ -1,3 +1,4 @@
 export { IsRootPageField } from '../components/server/IsRootPageField.js'
 export { ParentField } from '../components/server/ParentField.js'
+export { PathField } from '../components/server/PathField.js'
 export { SlugField } from '../components/server/SlugField.js'