+++ pages/src/fields/pathField.ts
@@ -19,7 +19,7 @@ export function pathField({ admin }: { admin?: TextField['admin'] } = {}): Field
     admin: mergeFieldAdmin<NonNullable<TextField['admin']>>(
       {
         components: {
-          Field: '@jhb.software/payload-pages-plugin/client#PathField',
+          Field: '@jhb.software/payload-pages-plugin/server#PathField',
         },
         disableBulkEdit: true,
         position: 'sidebar',